imputeTS: Time Series Missing Value Imputation

Imputation (replacement) of missing values in univariate time series. Offers several imputation functions and missing data plots. Available imputation algorithms include: 'Mean', 'LOCF', 'Interpolation', 'Moving Average', 'Seasonal Decomposition', 'Kalman Smoothing on Structural Time Series models', 'Kalman Smoothing on ARIMA models'. Published in Moritz and Bartz-Beielstein (2017) doi: 10.32614/RJ-2017-009 .

The imputeTS package is a collection of algorithms and tools for univariate time series imputation.

Details

The imputeTS package specializes on (univariate) time series imputation. It offers several different imputation algorithm implementations. Beyond the imputation algorithms the package also provides plotting and printing functions of missing data statistics.

The package is easy to use:

To impute (fill all missing values) in a time series x, run:
na_interpolation(x)

To plot missing data statistics for a time series x, run:
ggplot_na_distribution(x)

To print missing data statistics for a time series x, run:
statsNA(x)

Every other imputation function (starting with na_'algorithm name') and plotting function (starting with plotNA_'plot name') work the same way as in this example.

References

Moritz, Steffen, and Thomas Bartz-Beielstein. "imputeTS: Time Series Missing Value Imputation in R." R Journal 9.1 (2017). doi:10.32614/RJ-2017-009.

See also

Author

Maintainer: Steffen Moritz steffen.moritz10@gmail.com (ORCID) [copyright holder]

Authors:

Sebastian Gatscha sebastian_gatscha@gmx.at

Other contributors:

Earo Wang earo.wang@gmail.com (ORCID) [contributor]

Ron Hause ronaldhause@gmail.com (ORCID) [contributor]