Deprecated use na_kalman instead.

na.kalman is replaced by na_kalman. The functionality stays the same. The new name better fits modern R code style guidelines (which prefer _ over . in function names).

na.kalman(x, model = "StructTS", smooth = TRUE, nit = -1, maxgap = Inf,
  ...)

Arguments

x
Numeric Vector (vector) or Time Series (ts) object in which missing values shall be replaced

model
Model to be used. With this parameter the State Space Model (on which KalmanSmooth is performed) can be chosen. Accepts the following input:

"auto.arima" - For using the state space representation of arima model (using auto.arima)

"StructTS" - For using a structural model fitted by maximum likelihood (using StructTS)

For both auto.arima and StructTS additional parameters for model building can be given with the … parameter

Additionally it is also possible to use a user created state space model (See code Example 5). This state space model could for example be obtained from another R package for structural time series modeling. Furthermore providing the state space representation of a arima model from arima is also possible. But it is important to note, that user created state space models must meet the requirements specified under KalmanLike. This means the user supplied state space model has to be in form of a list with at least components T, Z, h , V, a, P, Pn. (more details under KalmanLike)

smooth
if TRUE - KalmanSmooth is used for estimation, if FALSE - KalmanRun is used. Since KalmanRun is often considered extrapolation KalmanSmooth is usually the better choice for imputation.

nit
Parameter from Kalman Filtering (see KalmanLike). Usually no need to change from default.

maxgap
Maximum number of successive NAs to still perform imputation on. Default setting is to replace all NAs without restrictions. With this option set, consecutive NAs runs, that are longer than 'maxgap' will be left NA. This option mostly makes sense if you want to treat long runs of NA afterwards separately.

...
Additional parameters to be passed through to the functions that build the State Space Models (StructTS or auto.arima).