Missing Value Imputation by Random Sample

Replaces each missing value by drawing a random sample between two given bounds.

na.random(x, lowerBound = min(x, na.rm = TRUE), upperBound = max(x, na.rm =
  TRUE))

Arguments

x
Numeric Vector (vector) or Time Series (ts) object in which missing values shall be replaced

lowerBound
Lower bound for the random samples (default is min(data) )

upperBound
Upper bound for the random samples (default is max(data) )

Value

Vector (vector) or Time Series (ts) object (dependent on given input at parameter x)

Details

Replaces each missing value by drawing a random sample between two given bounds. The default bounds are the minimum and the maximum value in the non-NAs from the time series. Function uses runif function to get the random values.

See also

Examples

#Prerequisite: Create Time series with missing values
x <- ts(c(2,3,NA,5,6,NA,7,8))

#Example 1: Replace all NAs by random values that are between min and max of the input time series
na.random(x)

#> Time Series:
#> Start = 1 
#> End = 8 
#> Frequency = 1 
#> [1] 2.000000 3.000000 2.484501 5.000000 6.000000 2.484501 7.000000 8.000000

#Example 2: Replace all NAs by random values between 1 and 10
na.random(x, lowerBound = 1, upperBound = 10)

#> Time Series:
#> Start = 1 
#> End = 8 
#> Frequency = 1 
#> [1] 2.000000 3.000000 8.508997 5.000000 6.000000 8.508997 7.000000 8.000000