Missing Value Imputation by Interpolation

Uses either linear, spline or stineman interpolation to replace missing values.

na_interpolation(x, option = "linear", maxgap = Inf, ...)

Arguments

x

Numeric Vector (vector) or Time Series (ts) object in which missing values shall be replaced

option

Algorithm to be used. Accepts the following input:

"linear" - for linear interpolation using approx (default choice)

"spline" - for spline interpolation using spline

"stine" - for Stineman interpolation using stinterp

maxgap

Maximum number of successive NAs to still perform imputation on. Default setting is to replace all NAs without restrictions. With this option set, consecutive NAs runs, that are longer than 'maxgap' will be left NA. This option mostly makes sense if you want to treat long runs of NA afterwards separately.

...

Additional parameters to be passed through to approx or spline interpolation functions

Value

Vector (vector) or Time Series (ts) object (dependent on given input at parameter x)

Details

Missing values get replaced by values of approx, spline or stinterp interpolation.

The na_interpolation function also supports the use of additional parameters from the respective underlying interpolation functions. While usually not really needed, it is useful to know that this advanced use is in principle possible. These additional parameters are not specified explicitly in the na_interpolation function documentation. Take a look into the documentation of the stinterp, approx and spline functions to get an overview about these additional parameters.

An example for such a parameter is the 'method' argument of spline, which can be used to further specify the type of spline to be used. Possible values are "fmm", "natural", "periodic", "monoH.FC" and "hyman" (as can be seen in the spline documentation). The respective function call using this additional parameter would look like this: na_interpolation(x, option ="spline", method ="natural")

Like in this example other additional detail parameters (gained from approx, spline, stinterp documentation) can be used by just including them in the na_interpolation function call. As already mentioned, these advanced possibilities for settings parameters are only helpful for specific use cases. For regular use the standard parameters provided directly in the na_interpolation documentation should be more than enough.

References

Johannesson, Tomas, et al. (2015). "Package stinepack".

See also

Author

Steffen Moritz, Ron Hause

Examples

# Prerequisite: Create Time series with missing values
x <- ts(c(2, 3, 4, 5, 6, NA, 7, 8))

# Example 1: Perform linear interpolation
na_interpolation(x)
#> Time Series:
#> Start = 1 
#> End = 8 
#> Frequency = 1 
#> [1] 2.0 3.0 4.0 5.0 6.0 6.5 7.0 8.0

# Example 2: Perform spline interpolation
na_interpolation(x, option = "spline")
#> Time Series:
#> Start = 1 
#> End = 8 
#> Frequency = 1 
#> [1] 2.00000 3.00000 4.00000 5.00000 6.00000 6.53923 7.00000 8.00000

# Example 3: Perform stine interpolation
na_interpolation(x, option = "stine")
#> Time Series:
#> Start = 1 
#> End = 8 
#> Frequency = 1 
#> [1] 2.0 3.0 4.0 5.0 6.0 6.5 7.0 8.0

# Example 4: Perform linear interpolation, with additional parameter pass through from spline()
# Take a look at the 'Details' section of the na_interpolation documentation 
# for more information about advanced parameter pass through options
na_interpolation(x, option ="spline", method ="natural")
#> Time Series:
#> Start = 1 
#> End = 8 
#> Frequency = 1 
#> [1] 2.00000 3.00000 4.00000 5.00000 6.00000 6.50661 7.00000 8.00000

# Example 5: Same as example 1, just written with pipe operator
x %>% na_interpolation()
#> Time Series:
#> Start = 1 
#> End = 8 
#> Frequency = 1 
#> [1] 2.0 3.0 4.0 5.0 6.0 6.5 7.0 8.0

# Example 6: Same as example 2, just written with pipe operator
x %>% na_interpolation(option = "spline")
#> Time Series:
#> Start = 1 
#> End = 8 
#> Frequency = 1 
#> [1] 2.00000 3.00000 4.00000 5.00000 6.00000 6.53923 7.00000 8.00000