Time series of monthly airline passengers (with NAs)

Monthly totals of international airline passengers, 1949 to 1960. This time series contains missing values. In the package included is also the tsAirgapComplete time series providing the true values for the missing values.

tsAirgap

Format

Time Series (ts) with 144 rows including 13 NAs.

Source

Box, G. E. P., Jenkins, G. M., Reinsel, G. C. and Ljung, G. M. (2015). Time series analysis: forecasting and control. Fifth Edition. John Wiley and Sons.

Details

The dataset originates from Box and Jenkins (see citation) and is a commonly used example in time series analysis literature.

It characteristics (strong trend, strong seasonal behavior) make it also a great example for time series imputation. Thus the version with inserted NA gaps was created under the name tsAirgap.

In order to use this series for comparing imputation algorithm results, there are two time series provided. One series without missing values, which can be used as ground truth. Another series with NAs, on which the imputation algorithms can be applied.

There are the two time series:

tsAirgap - The time series with NAs.

tsAirgapComplete - Time series without NAs.

See also