Deprecated use na_random instead.

na.random is replaced by na_random. The functionality stays the same. The new name better fits modern R code style guidelines (which prefer _ over . in function names).

na.random(x, lower_bound = NULL, upper_bound = NULL, maxgap = Inf, ...)

Arguments

x

Numeric Vector (vector) or Time Series (ts) object in which missing values shall be replaced

lower_bound

Lower bound for the random samples. If nothing or NULL is set min(x) will be used.

upper_bound

Upper bound for the random samples. If nothing or NULL is set man(x) will be used.

maxgap

Maximum number of successive NAs to still perform imputation on. Default setting is to replace all NAs without restrictions. With this option set, consecutive NAs runs, that are longer than 'maxgap' will be left NA. This option mostly makes sense if you want to treat long runs of NA afterwards separately.